o Added scan@dateFormatMark support and tests for #1444.
o This required adding a modified redistribution of the ICU project for handling SimpleDateFormat and formatting ISO 8601 dates.   It is built as part of the modules build process and designed to not interfere with other installations of ICU that may already exist.  We link the static libraries directly into the module shared lib to avoid dependency issues.